Refactor: message 모듈 파일 구조 개선 #52
