+++ app/src/main/AndroidManifest.xml
@@ -17,7 +17,7 @@
         tools:targetApi="31"
         android:usesCleartextTraffic="true"> <!-- HTTP 요청 허용 -->
         <activity
-            android:name=".MessageChatroomActivity"
+            android:name=".message.ui.MessageChatroomActivity"
             android:exported="false"
             android:windowSoftInputMode="adjustPan"/>
         <activity
