+++ app/src/main/java/com/example/hrr_android/message/adapter/MessageListRVAdapter.kt
@@ -1,10 +1,11 @@
-package com.example.hrr_android
+package com.example.hrr_android.message.adapter
 
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.example.hrr_android.databinding.ItemMessageListBinding
+import com.example.hrr_android.message.model.MessageList
 
 class MessageListRVAdapter(
     private val messageList: List<MessageList>,
