+++ app/src/main/java/com/example/hrr_android/message/model/MessageList.kt
@@ -1,4 +1,4 @@
-package com.example.hrr_android
+package com.example.hrr_android.message.model
 
 data class MessageList(
     val profileImg: Int,
