+++ app/src/main/java/com/example/hrr_android/message/ui/MessageFragment.kt
@@ -1,4 +1,4 @@
-package com.example.hrr_android
+package com.example.hrr_android.message.ui
 
 import android.content.Intent
 import android.os.Bundle
@@ -7,6 +7,9 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.LinearLayoutManager
+import com.example.hrr_android.message.model.MessageList
+import com.example.hrr_android.message.adapter.MessageListRVAdapter
+import com.example.hrr_android.R
 import com.example.hrr_android.databinding.FragmentMessageBinding
 
 class MessageFragment : Fragment() {
